requirements.txt
aiogram==3.22.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.1.1
